fix: moved related tests from WebSocket.test.ts to WebSocketServer.test.ts
